# perf: further speedup for 2.0 startup