--- a/MANIFEST.in
+++ b/MANIFEST.in
@@ -1,5 +1,6 @@
 recursive-include jina/resources *
 recursive-include jina/helloworld *
 include extra-requirements.txt
+include fastentrypoints.py
 prune tests/
 prune **/tests/