--- a/cli/api.py
+++ b/cli/api.py
@@ -88,7 +88,7 @@ def export_api(args: 'Namespace'):
     from .export import api_to_dict
     from jina.jaml import JAML
     from jina import __version__
-    from jina.logging import default_logger
+    from jina.logging.predefined import default_logger
     from jina.schemas import get_full_schema
 
     if args.yaml_path:
@@ -155,7 +155,7 @@ def flow(args: 'Namespace'):
         with f:
             f.block()
     else:
-        from jina.logging import default_logger
+        from jina.logging.predefined import default_logger
 
         default_logger.critical('start a flow from CLI requires a valid "--uses"')