--- a/jina/__init__.py
+++ b/jina/__init__.py
@@ -123,8 +123,6 @@ def _set_nofile(nofile_atleast=4096):
     except ImportError:  # Windows
         res = None
 
-    from .logging import default_logger
-
     if res is None:
         return (None,) * 2
 
@@ -136,21 +134,17 @@ def _set_nofile(nofile_atleast=4096):
         if hard < soft:
             hard = soft
 
-        default_logger.debug(f'setting soft & hard ulimit -n {soft} {hard}')
         try:
             res.setrlimit(res.RLIMIT_NOFILE, (soft, hard))
         except (ValueError, res.error):
             try:
                 hard = soft
-                default_logger.warning(
-                    f'trouble with max limit, retrying with soft,hard {soft},{hard}'
-                )
+                print(f'trouble with max limit, retrying with soft,hard {soft},{hard}')
                 res.setrlimit(res.RLIMIT_NOFILE, (soft, hard))
             except Exception:
-                default_logger.warning('failed to set ulimit, giving up')
+                print('failed to set ulimit, giving up')
                 soft, hard = res.getrlimit(res.RLIMIT_NOFILE)
 
-    default_logger.debug(f'ulimit -n soft,hard: {soft} {hard}')
     return soft, hard