--- a/jina/clients/request/asyncio.py
+++ b/jina/clients/request/asyncio.py
@@ -6,7 +6,7 @@ from .helper import _new_data_request_from_batch, _new_data_request
 from .. import GeneratorSourceType
 from ...enums import DataInputType
 from ...importer import ImportExtensions
-from ...logging import default_logger
+from ...logging.predefined import default_logger
 from ...types.request import Request