--- a/jina/helloworld/chatbot/app.py
+++ b/jina/helloworld/chatbot/app.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 from jina import Flow
 from jina.importer import ImportExtensions
-from jina.logging import default_logger
+from jina.logging.predefined import default_logger
 from jina.logging.profile import ProgressBar
 from jina.parsers.helloworld import set_hw_chatbot_parser
 from jina.types.document.generators import from_csv