--- a/jina/logging/__init__.py
+++ b/jina/logging/__init__.py
@@ -1,3 +1,0 @@
-from .logger import JinaLogger
-
-default_logger = JinaLogger('JINA')  #: a logger at the global-level