2.0rc1 already achieves 2.2x speedup on startup and `from jina import Document, Executor, Flow`. This PR further improves the speed up by carefully inspecting
```python
PYTHONPROFILEIMPORTTIME=x python -c 'from jina import Document, Flow, Executor'
```
and 
```python
PYTHONPROFILEIMPORTTIME=x jina -v
```

---
This PR achieves 1.6x speedup on `import` (3.6x compare to `1.x`):
```text
➜  jina git:(perf-startup) ✗ hyperfine --warmup 1 'python -c "from jina import Document, Flow, Executor"'
Benchmark #1: python -c "from jina import Document, Flow, Executor"
  Time (mean ± σ):     350.4 ms ±   6.1 ms    [User: 265.7 ms, System: 398.7 ms]
  Range (min … max):   343.2 ms … 363.2 ms    10 runs
```

while `master` is at:

```text
➜  jina git:(master) ✗ hyperfine --warmup 1 'python -c "from jina import Document, Flow, Executor"'
Benchmark #1: python -c "from jina import Document, Flow, Executor"
  Time (mean ± σ):     566.5 ms ±   7.3 ms    [User: 449.6 ms, System: 576.2 ms]
  Range (min … max):   556.3 ms … 582.8 ms    10 runs
```

---
This PR achieves 1.7x speedup on CLI `jina`  (3.4x compare to `1.x`)

```text
➜  jina git:(perf-startup) ✗ hyperfine --warmup 1 "jina -v"
Benchmark #1: jina -v
  Time (mean ± σ):     392.5 ms ±   3.6 ms    [User: 278.6 ms, System: 385.4 ms]
  Range (min … max):   386.3 ms … 397.0 ms    10 runs
```

while `master` is at:

```text
➜  jina git:(master) ✗ hyperfine --warmup 1 "jina -v"
Benchmark #1: jina -v
  Time (mean ± σ):     689.2 ms ±  14.3 ms    [User: 613.7 ms, System: 903.1 ms]
  Range (min … max):   673.2 ms … 712.7 ms    10 runs
```

Optimizations:
- remove the heavy usage of `pkg_resources ` (https://github.com/pypa/setuptools/issues/926, https://stackoverflow.com/a/9918496)
- carefully restructure/hide/remove head imports
- use https://github.com/ninjaaron/fast-entry_points to solve slow startup on Python 3.7.x (this is a bug in 3.7, 3.8+ does not require such patch)